requirements/base.txt
pytz==2022.1  # https://github.com/stub42/pytz
python-slugify==6.1.2  # https://github.com/un33k/python-slugify
Pillow==9.1.0  # https://github.com/python-pillow/Pillow
rcssmin==1.1.0  # https://github.com/ndparker/rcssmin
argon2-cffi==21.3.0  # https://github.com/hynek/argon2_cffi
whitenoise==6.0.0  # https://github.com/evansd/whitenoise
redis==4.2.2  # https://github.com/redis/redis-py
hiredis==2.0.0  # https://github.com/redis/hiredis-py
celery==5.2.6  # pyup: < 6.0  # https://github.com/celery/celery
django-celery-beat==2.2.1  # https://github.com/celery/django-celery-beat
flower==1.0.0  # https://github.com/mher/flower
uvicorn[standard]==0.17.6  # https://github.com/encode/uvicorn

# Django
# ------------------------------------------------------------------------------
django==3.2.13  # pyup: < 4.0  # https://www.djangoproject.com/
django-environ==0.8.1  # https://github.com/joke2k/django-environ
django-model-utils==4.2.0  # https://github.com/jazzband/django-model-utils
django-allauth==0.50.0  # https://github.com/pennersr/django-allauth
django-crispy-forms==1.14.0  # https://github.com/django-crispy-forms/django-crispy-forms
crispy-bootstrap5==0.6  # https://github.com/django-crispy-forms/crispy-bootstrap5
django-compressor==3.1  # https://github.com/django-compressor/django-compressor
django-redis==5.2.0  # https://github.com/jazzband/django-redis
django-postgres-extra==2.0.5 # https://github.com/SectorLabs/django-postgres-extra


factory-boy==3.2.1  # https://github.com/FactoryBoy/factory_boy
django-debug-toolbar==3.2.4  # https://github.com/jazzband/django-debug-toolbar
django-extensions==3.1.5  # https://github.com/django-extensions/django-extensions
django-coverage-plugin==2.0.2  # https://github.com/nedbat/django_coverage_plugin
pytest-django==4.5.2  # https://github.com/pytest-dev/pytest-django


# Django REST Framework
coreapi~=2.3.3
dj-rest-auth~=2.2.4
djangorestframework==3.13.1  # https://github.com/encode/django-rest-framework
django-cors-headers==3.11.0 # https://github.com/adamchainz/django-cors-headers
djangorestframework-jwt~=1.11.0
djangorestframework-simplejwt~=5.1.0
drf-nested-routers~=0.93.4
rest-condition~=1.0.3
djangorestframework-filters==1.0.0.dev0


# DRF yasg for api documentation
drf-yasg~=1.20.0
PyJWT==1.7.1

# Unicode Slugification
django-autoslug~=1.9.8

# python-slugify~=6.1.2
phonenumbers==8.12.48
twilio

# google cloud storages
google-cloud-storage==1.37.1
django-storages==1.12.3

docusign-esign
python-jose
cryptography

# pymongo
# pymongo==3.12.3
pymongo==4.3.3
djongo==1.3.6
dnspython==1.16.0
sqlparse==0.2.4
geopy==2.2.0
fcm-django==1.0.12
firebase-admin==5.2.0

# stripe payment gateway
stripe==3.5.0

httpx==0.23.1

django-filter==22.1

#search
uszipcode==1.0.1

